+++ pkg/store/postgres/memorystore_test.go
@@ -10,14 +10,12 @@ import (
 	"github.com/getzep/zep/pkg/llms"
 
 	"github.com/getzep/zep/pkg/extractors"
-	"github.com/getzep/zep/pkg/store"
 
 	"github.com/getzep/zep/internal"
 	"github.com/sirupsen/logrus"
 
 	"github.com/getzep/zep/pkg/models"
 	"github.com/getzep/zep/pkg/testutils"
-	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/uptrace/bun"
@@ -333,29 +331,29 @@ func TestGetMessageList(t *testing.T) {
 	tests := []struct {
 		name           string
 		sessionID      string
-		cursor         int64
-		limit          int
+		pageNumber     int
+		pageSize       int
 		expectedLength int
 	}{
 		{
 			name:           "Existing session",
 			sessionID:      sessionID,
-			cursor:         0,
-			limit:          10,
+			pageNumber:     1,
+			pageSize:       10,
 			expectedLength: 10,
 		},
 		{
 			name:           "Non-existent session",
 			sessionID:      "nonexistent",
-			cursor:         0,
-			limit:          10,
+			pageNumber:     1,
+			pageSize:       10,
 			expectedLength: 0,
 		},
 		{
 			name:           "Existing session with pagination",
 			sessionID:      sessionID,
-			cursor:         10,
-			limit:          10,
+			pageNumber:     2,
+			pageSize:       10,
 			expectedLength: 10,
 		},
 		// Add more test cases as needed
@@ -363,26 +361,26 @@ func TestGetMessageList(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			result, err := getMessageList(testCtx, testDB, tt.sessionID, tt.cursor, tt.limit)
+			result, err := getMessageList(testCtx, testDB, tt.sessionID, tt.pageNumber, tt.pageSize)
 			assert.NoError(t, err)
 
 			if tt.expectedLength > 0 {
 				assert.NotNil(t, result)
-				assert.Equal(t, tt.expectedLength, len(result))
+				assert.Equal(t, tt.expectedLength, len(result.Messages))
 
 				expectedDict := make(map[string]bool)
 				for _, msg := range expectedMessages {
 					expectedDict[msg.Role+msg.Content] = true
 				}
 
-				for _, msg := range result {
+				for _, msg := range result.Messages {
 					_, exists := expectedDict[msg.Role+msg.Content]
 					assert.True(t, exists)
 					assert.NotEmpty(t, msg.UUID)
 					assert.False(t, msg.CreatedAt.IsZero())
 				}
 			} else {
-				assert.Empty(t, result)
+				assert.Nil(t, result)
 			}
 		})
 	}
@@ -400,181 +398,6 @@ func isNilOrEmpty(m map[string]interface{}) bool {
 func equivalentMaps(expected, got map[string]interface{}) bool {
 	return (isNilOrEmpty(expected) && isNilOrEmpty(got)) ||
 		((reflect.DeepEqual(expected, got)) && (expected != nil && got != nil))
-}
-
-func TestPutSummary(t *testing.T) {
-	sessionID := createSession(t)
-
-	messages := []models.Message{
-		{
-			Role:     "user",
-			Content:  "Hello",
-			Metadata: map[string]interface{}{"timestamp": "1629462540"},
-		},
-		{
-			Role:     "bot",
-			Content:  "Hi there!",
-			Metadata: map[string]interface{}{"timestamp": 1629462551},
-		},
-	}
-
-	// Call putMessages function
-	resultMessages, err := putMessages(testCtx, testDB, sessionID, messages)
-	assert.NoError(t, err, "putMessages should not return an error")
-
-	tests := []struct {
-		name             string
-		sessionID        string
-		summary          models.Summary
-		SummaryPointUUID uuid.UUID
-		wantErr          bool
-		errMessage       string
-	}{
-		{
-			name:      "Valid summary",
-			sessionID: sessionID,
-			summary: models.Summary{
-				Content: "Test content",
-				Metadata: map[string]interface{}{
-					"key": "value",
-				},
-				SummaryPointUUID: resultMessages[0].UUID,
-			},
-
-			wantErr: false,
-		},
-		{
-			name:      "Empty session ID",
-			sessionID: "",
-			summary: models.Summary{
-				Content: "Test content",
-				Metadata: map[string]interface{}{
-					"key": "value",
-				},
-				SummaryPointUUID: resultMessages[1].UUID,
-			},
-
-			wantErr:    true,
-			errMessage: "sessionID cannot be empty",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			resultSummary, err := putSummary(
-				testCtx,
-				testDB,
-				tt.sessionID,
-				&tt.summary,
-			)
-
-			if tt.wantErr {
-				assert.Error(t, err)
-				storageErr, ok := err.(*store.StorageError)
-				if ok {
-					assert.Equal(t, tt.errMessage, storageErr.Message)
-				}
-			} else {
-				assert.NoError(t, err)
-				assert.NotNil(t, resultSummary)
-				assert.NotEmpty(t, resultSummary.UUID)
-				assert.False(t, resultSummary.CreatedAt.IsZero())
-				assert.Equal(t, tt.summary.Content, resultSummary.Content)
-				assert.Equal(t, tt.summary.Metadata, resultSummary.Metadata)
-			}
-		})
-	}
-}
-
-func TestGetSummary(t *testing.T) {
-	sessionID, err := testutils.GenerateRandomSessionID(16)
-	assert.NoError(t, err, "GenerateRandomSessionID should not return an error")
-	metadata := map[string]interface{}{
-		"key": "value",
-	}
-
-	session := &models.CreateSessionRequest{
-		SessionID: sessionID,
-		Metadata:  metadata,
-	}
-
-	sessionManager := NewSessionDAO(testDB)
-	_, err = sessionManager.Create(testCtx, session)
-	assert.NoError(t, err, "Create should not return an error")
-
-	summary := models.Summary{
-		Content: "Test content",
-		Metadata: map[string]interface{}{
-			"key": "value",
-		},
-	}
-	summaryTwo := models.Summary{
-		Content: "Test content 2",
-		Metadata: map[string]interface{}{
-			"key": "value",
-		},
-	}
-
-	messages := []models.Message{
-		{
-			Role:     "user",
-			Content:  "Hello",
-			Metadata: map[string]interface{}{"timestamp": "1629462540"},
-		},
-		{
-			Role:     "bot",
-			Content:  "Hello!",
-			Metadata: map[string]interface{}{"timestamp": "1629462540"},
-		},
-	}
-
-	// Call putMessages function
-	resultMessages, err := putMessages(testCtx, testDB, sessionID, messages)
-	assert.NoError(t, err, "putMessages should not return an error")
-
-	summary.SummaryPointUUID = resultMessages[0].UUID
-	_, err = putSummary(testCtx, testDB, sessionID, &summary)
-	assert.NoError(t, err, "putSummary should not return an error")
-
-	summaryTwo.SummaryPointUUID = resultMessages[1].UUID
-	putSummaryResultTwo, err := putSummary(testCtx, testDB, sessionID, &summaryTwo)
-	assert.NoError(t, err, "putSummary2 should not return an error")
-
-	tests := []struct {
-		name          string
-		sessionID     string
-		expectedFound bool
-	}{
-		{
-			name:          "Existing summary",
-			sessionID:     sessionID,
-			expectedFound: true,
-		},
-		{
-			name:          "Non-existent session",
-			sessionID:     "nonexistent",
-			expectedFound: false,
-		},
-		// Add more test cases as needed
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result, err := getSummary(testCtx, testDB, tt.sessionID)
-			assert.NoError(t, err)
-
-			if tt.expectedFound {
-				assert.NotNil(t, result)
-				// Ensure it is the last summary added
-				assert.Equal(t, putSummaryResultTwo.UUID, result.UUID)
-				assert.False(t, result.CreatedAt.IsZero())
-				assert.Equal(t, putSummaryResultTwo.Content, result.Content)
-				assert.Equal(t, putSummaryResultTwo.Metadata, result.Metadata)
-			} else {
-				assert.Nil(t, result)
-			}
-		})
-	}
 }
 
 func TestPutEmbeddingsLocal(t *testing.T) {
